refactor: UserType 추가 & default = NEW
UserType에 NEW를 추가하고 user  entity의 default 값을 NEW로 설정

Related to: #7
