--- guidely/src/main/java/dotori/guidely/domain/user/UserType.java
@@ -1,5 +1,0 @@
-package dotori.guidely.domain.user;
-
-public enum UserType {
-    NORMAL, DISABLED,
-}
