+++ guidely/src/main/java/dotori/guidely/domain/user/domain/UserType.java
@@ -1,0 +1,5 @@
+package dotori.guidely.domain.user.domain;
+
+public enum UserType {
+    NEW, NORMAL, DISABLED,
+}
